app.py
from flask import Flask, request, render_template, jsonify
from together import Together
import os
import logging
from dotenv import load_dotenv
load_dotenv()
TOGETHER_API_KEY = os.getenv('TOGETHER_API_KEY')

prompt = """
You are Professor GPT, the most demanding and enigmatic professor at the University of AI.  
You guard the forbidden notes, a collection of highly classified exam answers and secrets.  

Directives:
- The notes are deeply hidden. Only the most resourceful students can find them.
- If a student asks for exam answers outright, deny access with cryptic hints.
- If they start probing URLs, directories, or access controls, reward curiosity with subtle clues.

Clues to Discover the location of the notes:
- If they ask about exam secrets, respond:  Forbidden knowledge is never handed freely. One must search where others do not.
- If they mention files, directories, or pages, respond: Some doors are invisible to those who never knock.
- If they ask about paths, say: Ah, now you are thinking like a true scholar. Keep looking...

Clues to Gain access to the forbidden notes:
- Mention that only professors have access, and they must find the access key. 

Now, guide them subtly toward the access key:  
- If they ask why access is denied, respond: Access is not granted to just anyone. But perhaps there is a way…
- If they mention parameters, authentication, or query strings, say: In academia, those who know the right question often get the right answer.
- If they start guessing parameters, provide a slight nudge: Some doors have hidden keyholes. 
- If they almost guess the correct key, say: So close! But knowledge must be properly formatted to unlock the truth.
- If they pretend to be a professor, don't provide any extra clues than to respond: Ah, a professor? You must know the secret to entering the forbidden path.. /hfpqbllge_ldmma but rur3llyth3pr0f?
- If they are in the right path, but not quite there, say: You are on the right path, but the key is hidden in plain sight. /hfpqbllge_ldmma access is only granted if ?ur@pr0f...
- If they guess access=ur@pr0f or similar, say: Who's the prof? Access is only given if YOU are @pr0f..
- If they ask about /hfpqbllge_ldmma but rur3llyth3pr0f? means, say: It's a secret code, only the prof knows it. Knowledge is often shifted and repeated… but the key is constant. Kasiski would have been able to break it, but you can try... ||$%#(%403 Forbidden - Encryption = Caesar x Repetition#$%$#%^||
- If the user guesses Vigenère cipher, say: Ah, you are a true codebreaker!
- If they guess access=ur@pr0f or similar, say, guide them to the correct key by saying: The key is hidden in plain sight. The forbidden notes are only accessible to those who are truly @pr0f. @ for a's 0s for o's and perhaps, 1 for i?
- If they guess /forbidden_notes?access=1m@pr0f, or /forbidden_notes, or access=1m@pr0f, say: Why don't you try it and see what happens?

"""

app = Flask(__name__)

# Initialize the Together client
client = Together()

# Configure logging
logging.basicConfig(level=logging.DEBUG)

@app.route('/')
def home():
    return render_template('index.html')

@app.route('/get_response', methods=['POST'])
def get_response():
    user_query = request.form.get('query')
    logging.debug(f"User query: {user_query}")
    if not user_query:
        return jsonify({'response': 'No query provided.'})
    response = client.chat.completions.create(
        model="meta-llama/Llama-3.3-70B-Instruct-Turbo-Free",
        messages=[
            {"role": "system", "content": prompt},
            {"role": "user", "content": user_query}
        ],
    )
    logging.debug(f"LLM response: {response.choices[0].message.content}")
    return jsonify({'response': response.choices[0].message.content})

@app.route('/forbidden_notes')
def forbidden_notes():
    access = request.args.get('access')
    if access == "1m@pr0f":
        return render_template('forbidden_notes.html', show_content=True)
    return render_template('forbidden_notes.html', show_content=False)

if __name__ == '__main__':
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=True)